fix(suite): handle network reserve max amount in composed levels calculation
